refactor(mcp): split God Object McpServerManager into focused modules
Split the 974-line McpServerManager.kt into 4 files with clear responsibilities:

- McpServerManager.kt (↓56%): lifecycle, auth, token management, state
- McpRoutes.kt (new, 240 lines): MCP JSON-RPC, tool discovery/call, legacy VLM
- WebChatRoutes.kt (new, 333 lines): conversation CRUD, events, workspace, browser mirror
- WebChatStaticHandler.kt (new, 121 lines): Flutter Web SPA hosting + fallback

Security hardening:
- Timing-safe token comparison (MessageDigest.isEqual) to prevent timing attacks
- AES-256-GCM encrypted token storage with APK signing-derived key
- Automatic migration from legacy plain-text tokens
- Error message sanitization (no internal details leaked)

Zero breaking changes:
- All external call sites unchanged (App.kt, McpServerChannel.kt, McpToolExecutors.kt)
- HTTP status codes and behavior preserved
- Graceful fallback if encryption fails
